Move LambdaTypeName to TypeName.kt
